Added validation

+++ src/app/branches/branch-item/branch-item.component.ts
@@ -3,34 +3,28 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { BranchService } from '../branches.service';
 import { Branch } from '../branch.model';
 import { DataStorageService } from '../../shared/datastorage.service';
+import { ErrorService } from '../../error.service';
 
 @Component({
   selector: '[app-branch-item]',
   templateUrl: './branch-item.component.html',
   styleUrls: ['./branch-item.component.css']
 })
-export class BranchItemComponent implements OnInit {
+export class BranchItemComponent {
 
   @Input() branch: Branch;
-  @Input() index: number ;
+  @Input() index: number;
   constructor(private branchService: BranchService,
-    private router: Router, private datastorage: DataStorageService) {
-  }
-  ngOnInit() {
-    
-    //console.log(this.branch.buCode5);
-    //console.log(this.index);
-    //console.log(this.branch);
+    private router: Router, private datastorage: DataStorageService, private errorService: ErrorService) {
   }
   onEditBranch() {
-    console.log(this.branch.buCode5);
     this.router.navigate(['branches/', this.branch.buCode5, 'edit']);
   }
 
   onDeleteBranch() {
-
-    //this.branchService.deleteBranch(this.branch.buCode5);
-    this.datastorage.deleteBranch(this.branch.buCode5);
-    this.router.navigate(['/']);
+    if (confirm(`Do you want to delete branch ${this.branch.buCode5} ?`)) {
+      this.datastorage.deleteBranch(this.branch.buCode5);
+      this.router.navigate(['/']);
+    }
   }
 }
